+++ src/base/Field.php
@@ -138,6 +138,14 @@ abstract class Field extends SavableComponent implements FieldInterface
     /**
      * @inheritdoc
      */
+    public static function isRequirable(): bool
+    {
+        return true;
+    }
+
+    /**
+     * @inheritdoc
+     */
     public static function hasContentColumn(): bool
     {
         return true;